requirements.txt
streamlit
google-generativeai
pillow
python-dotenv
requests
openai
gTTS
pydub